Refactor data layer part 3
* Migrate fetchCheckoutHistory to XCircService
* Remove legacy ActorService, after migrating code
* Rename record.has* properties to align with Android
